Created BaseForm class, all form classes inherit from BaseForm for better management.
Created Method saveToProposal inside BaseForm, this method is overridden in all of the child classes, where the child classes will fill their data into the proposal object.

Created a method rememberProposal, the method is made generic and is called inside each of the posts to make an update to the proposal upon being called.

Created a class containing some static method to serve as a quick helper class and for some testing.